Update notification bar.
This introduces the update bar. If a new version is available, it can be downloaded or dismissed.
If the version is dismissed, the bar will not appear again for that version or any previous
versions.

Updates are checked every 24 hours, when the update is sucessful, or every 5 minutes until
success.

Fixes #16.