Update change in settings dismisses the ask bar.
If the user ignores the "do you want updates" bar and goes directly in to settings,
and enables "check for updates", the ask bar will remain visible. If "save" is used
in the settings dialog, then this now counts as "asking" if they want updates. This
seems safe since the default in the dialog is going to be "no" since we haven't asked
yet.

This also adds a preference watcher so the bar is automatically dismissed.