Removed 'dismissed' updates.
Now that there is a settings dialog to turn off updates completely,
this feels out-of-place and difficult to understand. This removes
the number of choices the user has to make when presented with an update.
If they really don't want to see the update after Fiddler restarts, they
can just turn the notifications off completely.